move handleSubscribedMembershipsList to a dir
