+++ client/state/data-layer/wpcom/sites/memberships/subscriptions/index.js
@@ -1,0 +1,38 @@
+/** @format */
+
+/**
+ * External dependencies
+ */
+
+import { noop } from 'lodash';
+
+/**
+ * Internal dependencies
+ */
+import {
+	MEMBERSHIPS_SUBSCRIPTIONS_LIST_REQUEST,
+	MEMBERSHIPS_SUBSCRIPTIONS_RECEIVE,
+} from 'state/action-types';
+
+import { http } from 'state/data-layer/wpcom-http/actions';
+import { dispatchRequestEx } from 'state/data-layer/wpcom-http/utils';
+
+export const handleSubscribedMembershipsList = dispatchRequestEx( {
+	fetch: action =>
+		http(
+			{
+				method: 'GET',
+				path: '/me/memberships/subscriptions',
+			},
+			action
+		),
+	onSuccess: ( action, { subscriptions, total } ) => ( {
+		type: MEMBERSHIPS_SUBSCRIPTIONS_RECEIVE,
+		subscriptions,
+		total,
+	} ),
+	onError: noop,
+} );
+export default {
+	[ MEMBERSHIPS_SUBSCRIPTIONS_LIST_REQUEST ]: [ handleSubscribedMembershipsList ],
+};
